[dev/height] seo article, refactoring,
