+++ next-sitemap.config.js
@@ -2,6 +2,7 @@ module.exports = {
   siteUrl: 'https://app.zziri.me',
   generateRobotsTxt: true,
   exclude: [
-    '**/_components/*',
+    // '**/_components/*',
+    '**/_*/*',
   ],
 };
